module reorganization, file cleaning
